R_MarkL1TextureCacheMRU_ asm

+++ m_near.c
@@ -234,12 +234,12 @@ int16_t             numtextures;
 
 int16_t activetexturepages[NUM_TEXTURE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
 uint8_t activenumpages[NUM_TEXTURE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
-int16_t textureL1LRU[NUM_TEXTURE_L1_CACHE_PAGES];
+uint8_t textureL1LRU[NUM_TEXTURE_L1_CACHE_PAGES];
 
 
 int16_t activespritepages[NUM_SPRITE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
 uint8_t activespritenumpages[NUM_SPRITE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
-int8_t  spriteL1LRU[NUM_SPRITE_L1_CACHE_PAGES];
+uint8_t spriteL1LRU[NUM_SPRITE_L1_CACHE_PAGES];
 
 int8_t spritecache_l2_head = -1;
 int8_t spritecache_l2_tail = -1;
